+++ superwall/src/main/java/com/superwall/sdk/logger/Logger.kt
@@ -1,3 +1,5 @@
+package com.superwall.sdk.logger
+
 import com.superwall.sdk.Superwall
 import com.superwall.sdk.config.options.SuperwallOptions